ideviceactivation: Check ActivationState for success if lockdownd_activate() returns no result
On older devices it might happen that lockdownd does not return any result,
neither success nor failure, when invoking lockdownd_activate(). This results
in the code to assume the operation failed.
Instead of relying on the return value of that function we query the actual
ActivationState to see if it is not "Unactivated" anymore to determine if the
activation was successful.